+++ integration_tests/twodut_oneotg_tests/bgp_triggered_gue_v6/bgp_triggered_gue_v6_test.go
@@ -35,7 +35,7 @@ import (
 	"github.com/openconfig/lemming/gnmi/fakedevice"
 	"github.com/openconfig/lemming/gnmi/oc"
 	"github.com/openconfig/lemming/gnmi/oc/ocpath"
-	"github.com/openconfig/lemming/integration_tests/binding"
+	"github.com/openconfig/lemming/internal/binding"
 	"github.com/openconfig/ondatra"
 	"github.com/openconfig/ondatra/gnmi"
 	otgtelemetry "github.com/openconfig/ondatra/gnmi/otg"
